Unified completion calculation and added _isOptional support
