+++ less/pageLevelProgress.less
@@ -37,6 +37,11 @@
     width:80%;
 }
 
+.page-level-progress-item-optional-text {
+    line-height: 1em;
+    text-align: right;
+}
+
 .page-level-progress-indicator {
     width: 16%;
     height: 10px;
